:heavy_plus_sign: Synchronize data after merge FinishedTasks
