+++ src/routes/historyPage/CalendarForm.scss
@@ -1,10 +1,8 @@
 .react-calendar {
-  width: 300px;
+  width: 280px;
   max-width: 100%;
   font-family: Arial, Helvetica, sans-serif;
   line-height: 1.125em;
-  border-radius: 8px;
-  box-shadow: 0 12px 24px rgba(0, 0, 0, 5%);
 }
 
 .react-calendar--doubleView {
