+++ src/routes/historyPage/HistoryPage.module.scss
@@ -25,7 +25,7 @@
   }
 
   h2 {
-    margin: 40px 0 20px;
+    margin: 20px 0 10px;
     font-size: sizes.$SUB_TITLE;
     color: colors.$SUB_TITLE;
   }
